LayoutTests/platform/gtk/accessibility/aria-labelledby-overrides-aria-labeledby-expected.txt

This tests that aria-labelledby overrides aria-labeledby correctly.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS linkTitle is 'AXTitle: Y Z'
PASS successfullyParsed is true

TEST COMPLETE
